feat(ModelOverrideMenu): implement new component for model selection in dialog view; refactor DialogView to utilize the new component
